requirements.txt
fastapi
pydantic
psycopg2-binary
uvicorn
